backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Anton"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.4"
Pillow = "^8.2.0"
djangorestframework = "^3.12.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
